Use quotes around paths and mkdir -p in backup/restore
Keeping all paths and variables in quotes is a good practice to avoid
problems when the paths and variables contain special symbols (e.g.
spaces).

mkdir -p is simpler than multiple mkdir calls and if statements. We
don't check if mkdir was successful either way, so it's better to just
use mkdir -p here.